+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **New Feature**: You now have more fine-grained control over how your files
   are displayed: You can now select if the filename is always used or a title
   or first heading level 1 is used (if applicable)
+- **New Feature**: You can now also fold nested lists
 - Fixed an error that would cause the global search to malfunction if you
   followed a link which then started a global search while the file manager was
   hidden